% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.cabin.R
\name{build.cabin}
\alias{build.cabin}
\title{Build Cabin}
\usage{
build.cabin()
}
\description{
Builds a warm house.
}
\details{
Costs 20 building materials and 6 construction points. Provides a home which will keep you warm on approximately two thirds of nights.
}
\seealso{
Other buildings: \code{\link{build.survivalshelter}},
  \code{\link{view.buildings}}
}
