% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.survivalshelter.R
\name{build.survivalshelter}
\alias{build.survivalshelter}
\title{Build Survival Shelter}
\usage{
build.survivalshelter()
}
\description{
Builds a simple survival shelter.
}
\details{
Costs 2 building materials and 1 construction point. Provides a little shelter which will keep you warm on approximately a third of nights.
}
\seealso{
Other buildings: \code{\link{build.cabin}},
  \code{\link{view.buildings}}
}
