% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean}
\usage{
clean()
}
\description{
Cleans, preventing squalor gain and decreasing squalor. For non-daily or monthly cleaning.
}
\seealso{
Other activities: \code{\link{did.daily}},
  \code{\link{did.monthly}}, \code{\link{did.weekly}},
  \code{\link{scavenge}}
}

