% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\title{Default Tasks}
\usage{
defaults(name, enabled = FALSE)
}
\arguments{
\item{name}{Specifies which default task will be changed.}

\item{enabled}{Defaults to \code{FALSE}, which disables the task. \code{enabled = TRUE} allows you to re-enable previously disabled defaults.}
}
\description{
Disables / Enables default dailies, monthlies, and weeklies.
}
\seealso{
Other task configuration: \code{\link{delete.daily}},
  \code{\link{delete.monthly}},
  \code{\link{delete.weekly}}, \code{\link{new.daily}},
  \code{\link{new.monthly}}, \code{\link{new.weekly}}
}
