% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.daily.R
\name{delete.daily}
\alias{delete.daily}
\title{Delete daily}
\usage{
delete.daily(name, completed = FALSE)
}
\arguments{
\item{name}{determines which daily will be deleted. Must be a string.}

\item{completed}{should be set to TRUE if the daily was already already completed this day. Defaults to FALSE.}
}
\description{
Deletes an existing custom daily.
}
\seealso{
Other task configuration: \code{\link{defaults}},
  \code{\link{delete.monthly}},
  \code{\link{delete.weekly}}, \code{\link{new.daily}},
  \code{\link{new.monthly}}, \code{\link{new.weekly}}
}
