% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.weekly.R
\name{delete.weekly}
\alias{delete.weekly}
\title{Delete weekly}
\usage{
delete.weekly(name, completed = FALSE)
}
\arguments{
\item{name}{determines which weekly will be deleted. Must be a string.}

\item{completed}{should be set to TRUE if the weekly was already already completed this day. Defaults to FALSE.}
}
\description{
Deletes an existing custom weekly.
}
\seealso{
Other task configuration: \code{\link{defaults}},
  \code{\link{delete.daily}}, \code{\link{delete.monthly}},
  \code{\link{new.daily}}, \code{\link{new.monthly}},
  \code{\link{new.weekly}}
}
