% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival game.R
\name{did.daily}
\alias{did.daily}
\title{Did Daily}
\usage{
did.daily(name, x = 15)
}
\arguments{
\item{name}{determines which task will be marked as complete.}

\item{x}{Defaults to 15. Sets how long the time was for scavenging in walk-type dailies.}
}
\description{
Checks of a daily task as done.
}
\seealso{
Other activities: \code{\link{clean}},
  \code{\link{did.monthly}}, \code{\link{did.weekly}},
  \code{\link{scavenge}}
}
