% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.monthly.R
\name{did.monthly}
\alias{did.monthly}
\title{Did Monthly}
\usage{
did.monthly(name)
}
\arguments{
\item{name}{determines which task will be marked as complete.}
}
\description{
Checks of a monthly task as done.
}
\seealso{
Other activities: \code{\link{clean}},
  \code{\link{did.daily}}, \code{\link{did.weekly}},
  \code{\link{scavenge}}
}

