% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.weekly.R
\name{did.weekly}
\alias{did.weekly}
\title{Did Weekly}
\usage{
did.weekly(name)
}
\arguments{
\item{name}{determines which task will be marked as complete.}
}
\description{
Checks of a weekly task as done.
}
\seealso{
Other activities: \code{\link{clean}},
  \code{\link{did.daily}}, \code{\link{did.monthly}},
  \code{\link{scavenge}}
}

