% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival game.R
\name{learn.lesserheal}
\alias{learn.lesserheal}
\title{Learn Lesser Heal}
\usage{
learn.lesserheal()
}
\description{
Spends a level up to learn the spell Lesser Heal.
}
\details{
Requires one mana to learn and cast. Heals one point of health.
}
