% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival game.R
\name{new.game}
\alias{new.game}
\title{New Game}
\usage{
new.game(location, day = "monday")
}
\arguments{
\item{location}{Sets the save file location and name, e.g. "C:/Users/.../Survival Game/survivalsave.csv". Must be .csv format.}

\item{day}{Sets the current day for use in \code{\link{run.day}}. Must be in lowercase. Defaults to "monday"}
}
\description{
Starts a new game, saved at the specified location.
}
\details{
This function starts a new game, setting all values to the default and saving them to the directory and file specified by the location argument.
}
