% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.monthly.R
\name{new.monthly}
\alias{new.monthly}
\title{New Monthly}
\usage{
new.monthly(name)
}
\arguments{
\item{name}{sets the name of the task. Must be a short string.}
}
\description{
Creates a new monthly task
}
\seealso{
Other task configuration: \code{\link{defaults}},
  \code{\link{delete.daily}}, \code{\link{delete.monthly}},
  \code{\link{delete.weekly}}, \code{\link{new.daily}},
  \code{\link{new.weekly}}
}
