% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.weekly.R
\name{new.weekly}
\alias{new.weekly}
\title{New Weekly}
\usage{
new.weekly(name)
}
\arguments{
\item{name}{sets the name of the task. Must be a short string.}
}
\description{
Creates a new weekly task
}
\seealso{
Other task configuration: \code{\link{defaults}},
  \code{\link{delete.daily}}, \code{\link{delete.monthly}},
  \code{\link{delete.weekly}}, \code{\link{new.daily}},
  \code{\link{new.monthly}}
}
