% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival game.R
\name{run.day}
\alias{run.day}
\title{Run Day}
\usage{
run.day(endmonth = FALSE)
}
\arguments{
\item{endmonth}{defaults to \code{FALSE} To reset montlies use the \code{endmonth = TRUE}.}
}
\description{
Run the day and process stat changes.
}
\details{
Runs the day and updates all the stats that change daily. Should be done daily, as it updates the day of the week. Resets dailies, and weeklies when run on Sunday. 
Be sure to only do this after you're done all of the things you wanted to do that day, because you can take damage! Does not automatically track months.
}
