% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival game.R
\name{save.game}
\alias{save.game}
\title{Save Game}
\usage{
save.game(location = savespot)
}
\description{
Saves the game to the specified location.
}
\details{
Only necessary if you're saving to a different file than your default one set by \code{\link{load.game}} or \code{\link{new.game}}. Does not change autosave location, the game must be reloaded to do so.
}
