% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival game.R
\name{scavenge}
\alias{scavenge}
\title{Scavenge}
\usage{
scavenge(x = 15)
}
\description{
Scavenges for supplies when going on a walk. For non-daily walks.
}
\details{
Takes an argument representing minutes walking, for example \code{did.walk(15)}. Defaults to 15. Each 15 minute incrament provides a bonus to scavenging.
}
\seealso{
Other activities: \code{\link{clean}},
  \code{\link{did.daily}}, \code{\link{did.monthly}},
  \code{\link{did.weekly}}
}
