% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.buildings.R
\name{view.buildings}
\alias{view.buildings}
\title{View Buildings}
\usage{
view.buildings(available = FALSE)
}
\arguments{
\item{available}{Displays building options rather than already constructed. Defaults to \code{FALSE}.}
}
\description{
Displays a list of buildings constructed or available for construction.
}
\seealso{
Other display: \code{\link{view.character}},
  \code{\link{view.tasks}}

Other buildings: \code{\link{build.cabin}},
  \code{\link{build.survivalshelter}}
}
