% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.character.R
\name{view.character}
\alias{view.character}
\title{View Character}
\usage{
view.character()
}
\description{
Displays information about your character's stats and essential supplies.
}
\seealso{
Other display: \code{\link{view.tasks}}
}
