% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.spells.R
\name{view.spells}
\alias{view.spells}
\title{View Spells}
\usage{
view.spells(available = FALSE)
}
\arguments{
\item{available}{Displays spell options rather than already known. Defaults to \code{FALSE}.}
}
\description{
Displays a list of spells known or available to be learned.
}
\seealso{
Other display: \code{\link{view.buildings}},
  \code{\link{view.character}}, \code{\link{view.tasks}}

Other spells: \code{\link{cast.lesserheal}},
  \code{\link{learn.lesserheal}}
}
