% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.tasks.R
\name{view.tasks}
\alias{view.tasks}
\title{View Tasks}
\usage{
view.tasks()
}
\description{
Displays tasks and if they have been completed or not.
}
\seealso{
Other display: \code{\link{view.buildings}},
  \code{\link{view.character}}, \code{\link{view.spells}}
}
