% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.game.R
\name{load.game}
\alias{load.game}
\title{Load Game}
\usage{
load.game(location)
}
\arguments{
\item{location}{Sets the save file location and name, e.g. "C:/Users/.../Survival Game/survivalsave.csv". Must be .csv format.}
}
\description{
Loads the game from the specified file.
}
\details{
Loads the game from the specified file and sets that as the save spot for all updates. Note that the game auto-saves whenever changes are made.
}

