% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.daily.R
\name{new.daily}
\alias{new.daily}
\title{New Daily}
\usage{
new.daily(name, type)
}
\arguments{
\item{name}{sets the name of the task. Must be a short string.}

\item{type}{sets which type of reward is granted for completing the daily. Valid types are "walk", "clean", "mana", "none"}
}
\description{
Creates a new daily task
}
\details{
Walk-type dailies apply the \code{\link{scavenge}} action automatically when completed. Clean-type dailies apply the \code{\link{clean}} activity automatically when completed. Mana-type dailies raise mana by one point when completed.
}
\seealso{
Other task configuration: \code{\link{defaults}},
  \code{\link{delete.daily}}, \code{\link{delete.monthly}},
  \code{\link{delete.weekly}}, \code{\link{new.monthly}},
  \code{\link{new.weekly}}
}
